[Quant][devs] Separated implementations for quantized & non-quantized tensors for the squeeze function
Summary: This PR is part of a series of PRs addressing https://github.com/pytorch/pytorch/issues/54150,
related to using dispatcher for calls to quantized backends as opposed to if/else conditionals.
This particular PR separates the calls to quantized & non-quantized backends for squeeze
using a dispatcher.

Test plan: Additional testing was not implemented to test this change because test cases in the existing test case already make use of the squeeze function for various backends.

ghstack-source-id: 01c5bc3cefc1e14f0d1c304e8a1b291ceb7205a9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71639

initial