[Quant][devs] Separated implementations for quantized & non-quantized tensors for the unsqueeze function
Summary: This PR is part of a series of PRs addressing https://github.com/pytorch/pytorch/issues/54150,
related to using dispatcher for calls to quantized backends as opposed to if/else conditionals.
This particular PR separates the calls to quantized & non-quantized backends for unsqueeze
using a dispatcher.

Test plan: Additional testing was not implemented to test this change because test cases in the existing test suite already make use of the squeeze function for various backends.

ghstack-source-id: 1b5c4ad86c66c6ae7fe1bd4c8f84d687d04bb15c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71648