[Quant][devs] Changed empty_quantized call for quantized tensor to resize_output
Summary: This PR is part of a series of PRs addressing https://github.com/pytorch/pytorch/issues/54150,
related to using dispatcher for calls to quantized backends as opposed to if/else conditionals.
This particular PR removes the call to empty_quantized for quantized tensors and substitutes
it for resize_output, which works for quantized tensors, based on current understanding.
Using the dispatcher for this function was determined to be not practical as it would entail
a significant amoutn of duplicate code

ghstack-source-id: b1b1516dbb97954e3a960b881fc10b613d1b8b2f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71899