[Quant][devs] Added early return statement in squeeze_qtensor when input tensor has zero dimensions or the input dim is not size 1
Summary: Summary: This PR is part of a series of PRs addressing https://github.com/pytorch/pytorch/issues/54150,
related to using dispatcher for calls to quantized backends as opposed to if/else conditionals.
This particular PR doesn't address any dispatcher issues but is the first of 2 stacked PRs that addresses separating
the implementations for quantized & non-quantized squeeze functions.

ghstack-source-id: 1d6c859547aabd61feb94061662dc6d189880187
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71876